[launch] increase sleep to .1
Fixes a bug with dwm. When no window is open and launch is used to open
a new window then the new window doesn't get focused for some reason. It
might be a bug with dwm but it gets fixed when the launch window doesn't
get closed that fast. It seems to happen only when st is opened and not
firefox. Maybe st is opened to fast.

This also removes flickering when opening a new terminal in dwm, because
otherwise the window below the launcher (usually the browser in my case)
is visible for a brief period of time after the launcher closes and
before the new terminal opens.